requirements.txt
# Core ML
catboost>=1.2
xgboost>=2.0
lightgbm>=4.0
scikit-learn>=1.3

# Optimization
optuna>=3.4

# Data
pandas>=2.0
numpy>=1.24

# LLM Integration
anthropic>=0.25
openai>=1.0
python-dotenv>=1.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Scientific
scipy>=1.11

# Testing
pytest>=7.0
